consolidate styling
